+++ LICENSE
@@ -222,7 +222,7 @@ core/src/main/resources/org/apache/spark/ui/static/d3-flamegraph.css
 Python Software Foundation License
 ----------------------------------
 
-python/pyspark/loose_version.py
+python/pyspark/sql/loose_version.py
 
 BSD 3-Clause
 ------------
