+++ LICENSE-binary
@@ -424,7 +424,7 @@ for text of these licenses.
 Python Software Foundation License
 ----------------------------------
 
-python/pyspark/loose_version.py
+python/pyspark/sql/loose_version.py
 
 
 BSD 2-Clause
