+++ python/pyspark/pandas/plot/matplotlib.py
@@ -15,7 +15,7 @@
 # limitations under the License.
 #
 
-from pyspark.loose_version import LooseVersion
+from pyspark.sql.loose_version import LooseVersion
 
 import matplotlib as mat
 import numpy as np
