+++ python/pyspark/pandas/tests/diff_frames_ops/test_corrwith.py
@@ -19,7 +19,7 @@ import pandas as pd
 import numpy as np
 
 from pyspark import pandas as ps
-from pyspark.loose_version import LooseVersion
+from pyspark.sql.loose_version import LooseVersion
 from pyspark.pandas.config import set_option, reset_option
 from pyspark.testing.pandasutils import PandasOnSparkTestCase
 from pyspark.testing.sqlutils import SQLTestUtils
