+++ python/pyspark/pandas/tests/indexes/test_symmetric_diff.py
@@ -20,7 +20,7 @@ import unittest
 import pandas as pd
 
 import pyspark.pandas as ps
-from pyspark.loose_version import LooseVersion
+from pyspark.sql.loose_version import LooseVersion
 from pyspark.testing.pandasutils import PandasOnSparkTestCase, TestUtils
