+++ python/pyspark/sql/connect/utils.py
@@ -16,7 +16,7 @@
 #
 import sys
 
-from pyspark.loose_version import LooseVersion
+from pyspark.sql.loose_version import LooseVersion
 from pyspark.sql.pandas.utils import require_minimum_pandas_version, require_minimum_pyarrow_version
 from pyspark.errors import PySparkImportError
