+++ python/pyspark/sql/pandas/utils.py
@@ -15,7 +15,7 @@
 # limitations under the License.
 #
 
-from pyspark.loose_version import LooseVersion
+from pyspark.sql.loose_version import LooseVersion
 from pyspark.errors import PySparkImportError, PySparkRuntimeError
