+++ python/pyspark/tests/test_util.py
@@ -21,7 +21,6 @@ from py4j.protocol import Py4JJavaError
 
 from pyspark import keyword_only
 from pyspark.util import _parse_memory
-from pyspark.loose_version import LooseVersion
 from pyspark.testing.utils import PySparkTestCase, eventually
 from pyspark.find_spark_home import _find_spark_home
 
@@ -106,18 +105,6 @@ class UtilTests(PySparkTestCase):
         eventually(timeout=180, catch_assertions=True)(
             lambda: self.assertTrue(random.random() < 0.1)
         )()
-
-    def test_loose_version(self):
-        v1 = LooseVersion("1.2.3")
-        self.assertEqual(str(v1), "1.2.3")
-        self.assertEqual(repr(v1), "LooseVersion ('1.2.3')")
-        v2 = "1.2.3"
-        self.assertTrue(v1 == v2)
-        v3 = 1.1
-        with self.assertRaises(TypeError):
-            v1 > v3
-        v4 = LooseVersion("1.2.4")
-        self.assertTrue(v1 <= v4)
 
     def test_parse_memory(self):
         self.assertEqual(_parse_memory("1g"), 1024)
